requirements/base.txt
# Django and REST framework
django==4.2.7
djangorestframework==3.14.0

# FastAPI and dependencies
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.4.2
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
httpx==0.24.1

# Database
psycopg2-binary==2.9.7
sqlalchemy==2.0.21

# File processing
pandas==2.1.1
numpy==1.26.0
openpyxl==3.1.2
PyPDF2==3.0.1
pyarrow==13.0.0
fastparquet==2023.8.0

# Testing
pytest==7.4.2
pytest-django==4.5.2

# Utilities
python-dotenv==1.0.0
pydantic-settings==2.0.3